requirements.txt
requests>=2.25.1
flask>=2.3.0
python-dotenv>=1.0.0
